+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 - fixed airtable key verification
 - improved HTML rendering (hash and ecli endpoints)
 - added list endpoint for ecli navigation
+- added document language attribute
+- DB Schema Update `add_lang.sql`
 
 
 ### [0.1.0] - (2020-11-13)